Move code to better place, add unit test
