+++ superset/assets/src/modules/utils.js
@@ -3,6 +3,7 @@ import d3 from 'd3';
 import $ from 'jquery';
 
 import { formatDate, UTC } from './dates';
+import { TIME_SHIFT_PATTERN } from '../utils/common';
 
 const siFormatter = d3.format('.3s');
 
@@ -275,3 +276,44 @@ export function mainMetric(savedMetrics) {
   }
   return metric;
 }
+
+/**
+ * Get a CSS class for styling a time series.
+ *
+ * The main time series should be styled as 'solid', and different time shifts
+ * are assigned different sytles (dashed, dotted, dash-dotted, etc.).
+ *
+ * The function cycles through ten different styles while memoizing offsets
+ * association to classes. If the function is called twice with the
+ * same offset it will return the same class.
+ *
+ * @param {string} key - The time series key for which we want to get a class
+ */
+export const getClassed = (function () {
+  const seen = {};
+  let i = 0;
+  return function (key) {
+    if (!key || !Array.isArray(key)) {
+      return 'solid';
+    }
+
+    const classes = [];
+    for (let j = 0; j < 10; j++) {
+      classes.push('time-shift-' + j);
+    }
+
+    let offset = null;
+    key.forEach((k) => {
+      if (TIME_SHIFT_PATTERN.test(k)) {
+        offset = k;
+      }
+    });
+    if (offset === null) {
+      return 'solid';
+    }
+    if (seen[offset] === undefined) {
+      seen[offset] = classes[i++ % classes.length];
+    }
+    return seen[offset];
+  };
+}());
